refactor: 마이페이지 관련 코드 분리
