+++ src/main/java/com/chooz/post/application/PostQueryService.java
@@ -4,23 +4,19 @@ import com.chooz.comment.domain.CommentRepository;
 import com.chooz.common.dto.CursorBasePaginatedResponse;
 import com.chooz.common.exception.BadRequestException;
 import com.chooz.common.exception.ErrorCode;
-import com.chooz.post.application.dto.PollChoiceVoteInfo;
 import com.chooz.post.application.dto.PostWithVoteCount;
 import com.chooz.post.domain.PollChoice;
-import com.chooz.post.domain.PollChoiceRepository;
 import com.chooz.post.domain.Post;
 import com.chooz.post.domain.PostRepository;
 import com.chooz.post.presentation.UpdatePostResponse;
 import com.chooz.post.presentation.dto.AuthorDto;
 import com.chooz.post.application.dto.FeedDto;
 import com.chooz.post.presentation.dto.FeedResponse;
-import com.chooz.post.presentation.dto.MostVotedPollChoiceDto;
 import com.chooz.post.presentation.dto.MyPagePostResponse;
 import com.chooz.post.presentation.dto.PollChoiceVoteResponse;
 import com.chooz.post.presentation.dto.PostResponse;
 import com.chooz.user.domain.User;
 import com.chooz.user.domain.UserRepository;
-import com.chooz.vote.application.RatioCalculator;
 import com.chooz.vote.domain.Vote;
 import com.chooz.vote.domain.VoteRepository;
 import lombok.RequiredArgsConstructor;
@@ -31,11 +27,7 @@ import org.springframework.data.domain.SliceImpl;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Comparator;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 
 @Service
 @Transactional(readOnly = true)
@@ -43,11 +35,10 @@ import java.util.stream.Collectors;
 public class PostQueryService {
 
     private final PostRepository postRepository;
-    private final PollChoiceRepository pollChoiceRepository;
     private final UserRepository userRepository;
     private final VoteRepository voteRepository;
     private final CommentRepository commentRepository;
-    private final RatioCalculator ratioCalculator;
+    private final MyPagePostManager myPagePostManager;
 
     public PostResponse findByShareUrl(Long userId, String shareUrl) {
         Post post = postRepository.findByShareUrlFetchPollChoices(shareUrl)
@@ -101,72 +92,11 @@ public class PostQueryService {
     }
 
     public CursorBasePaginatedResponse<MyPagePostResponse> findUserPosts(Long userId, Long cursor, int size) {
-        Slice<PostWithVoteCount> postSlice = postRepository.findPostsWithVoteCountByUserId(userId, cursor, Pageable.ofSize(size));
-
-        return getCursorPaginatedResponse(postSlice);
+        return myPagePostManager.getUserPosts(userId, cursor, Pageable.ofSize(size));
     }
 
     public CursorBasePaginatedResponse<MyPagePostResponse> findVotedPosts(Long userId, Long cursor, int size) {
-        Slice<PostWithVoteCount> postSlice = postRepository.findVotedPostsWithVoteCount(userId, cursor, Pageable.ofSize(size));
-
-        return getCursorPaginatedResponse(postSlice);
-    }
-
-    private CursorBasePaginatedResponse<MyPagePostResponse> getCursorPaginatedResponse(Slice<PostWithVoteCount> postSlice) {
-        if (postSlice.isEmpty()) {
-            return CursorBasePaginatedResponse.of(new SliceImpl<>(
-                    List.of(),
-                    postSlice.getPageable(),
-                    false
-            ));
-        }
-
-        Map<Long, PollChoiceVoteInfo> mostVotedPollChoiceByPostId = getMostVotedPollChoiceByPostId(getPostIds(postSlice));
-
-        List<MyPagePostResponse> response = getMyPagePostResponses(postSlice, mostVotedPollChoiceByPostId);
-
-        return CursorBasePaginatedResponse.of(new SliceImpl<>(
-                response,
-                postSlice.getPageable(),
-                postSlice.hasNext()
-        ));
-    }
-
-    private List<MyPagePostResponse> getMyPagePostResponses(
-            Slice<PostWithVoteCount> postSlice,
-            Map<Long, PollChoiceVoteInfo> mostVotedPollChoiceByPostId
-    ) {
-        return postSlice.getContent().stream()
-                .map(postWithVoteCount -> {
-                    var pollChoiceVoteInfo = mostVotedPollChoiceByPostId.get(postWithVoteCount.post().getId());
-                    var mostVotedPollChoiceInfo = MostVotedPollChoiceDto.of(
-                            pollChoiceVoteInfo,
-                            ratioCalculator.calculate(postWithVoteCount.voteCount(), pollChoiceVoteInfo.voteCounts())
-                    );
-                    return MyPagePostResponse.of(postWithVoteCount, mostVotedPollChoiceInfo);
-                })
-                .toList();
-    }
-
-    private Map<Long, PollChoiceVoteInfo> getMostVotedPollChoiceByPostId(List<Long> postIds) {
-        List<PollChoiceVoteInfo> pollChoiceWithVoteInfo = pollChoiceRepository.findPollChoiceWithVoteInfo(postIds);
-        return pollChoiceWithVoteInfo.stream()
-                .collect(Collectors.groupingBy(
-                        PollChoiceVoteInfo::postId,
-                        Collectors.collectingAndThen(
-                                Collectors.toList(),
-                                choices -> choices.stream()
-                                        .max(Comparator.comparing(PollChoiceVoteInfo::voteCounts))
-                                        .orElse(null)
-                        )
-                ));
-    }
-
-    private List<Long> getPostIds(Slice<PostWithVoteCount> postSlice) {
-        return postSlice.getContent()
-                .stream()
-                .map(postWithVoteCount -> postWithVoteCount.post().getId())
-                .toList();
+        return myPagePostManager.getVotedPosts(userId, cursor, Pageable.ofSize(size));
     }
 
     public CursorBasePaginatedResponse<FeedResponse> findFeed(Long userId, Long cursor, int size) {
